Fixed python scripts to avoid throwing an error on first run.

+++ facts.d/get_directives_details.py
@@ -1,26 +1,29 @@
 #!/usr/bin/python
 
 import json
+import os.path
 import re
 import subprocess
 
 zs_api_config_file = '/.zsapi.ini'
 zs_api_target = 'localadmin'
 
-directives_details = subprocess.check_output(["/usr/local/zend/bin/zs-client.sh", "configurationDirectivesList", "--target=localadmin", "--output-format=json"])
+if os.path.isfile("/usr/local/zend/bin/zs-client.sh"):
 
-## Strip the PHP notices from the json
-directives_details = re.sub("Notice:.*\n", "", directives_details)
-## Strip the newlines from the json
-directives_details = re.sub("\n", "", directives_details)
+  directives_details = subprocess.check_output(["/usr/local/zend/bin/zs-client.sh", "configurationDirectivesList", "--target=localadmin", "--output-format=json"])
 
-f1=open('/tmp/puppet_zend_directives.txt', 'w')
-f1.write(directives_details)
+  ## Strip the PHP notices from the json
+  directives_details = re.sub("Notice:.*\n", "", directives_details)
+  ## Strip the newlines from the json
+  directives_details = re.sub("\n", "", directives_details)
 
-arr = json.loads(directives_details)
+  f1=open('/tmp/puppet_zend_directives.txt', 'w')
+  f1.write(directives_details)
 
-for directive in arr[u"responseData"][u"directives"]:
-  name = directive["name"]
-  for key, value in directive.iteritems():
-    if value and not isinstance(value, list) and not isinstance(value,dict):
-      print ('zend_directive_' + key + '_' + name + '=' + value)
+  arr = json.loads(directives_details)
+
+  for directive in arr[u"responseData"][u"directives"]:
+    name = directive["name"]
+    for key, value in directive.iteritems():
+      if value and not isinstance(value, list) and not isinstance(value,dict):
+        print ('zend_directive_' + key + '_' + name + '=' + value)
